+++ src/main/java/kitchenpos/products/domain/Product.java
@@ -1,7 +1,5 @@
 package kitchenpos.products.domain;
 
-import kitchenpos.products.infra.PurgomalumClient;
-
 import javax.persistence.Column;
 import javax.persistence.Embedded;
 import javax.persistence.Entity;
