+++ src/main/java/kitchenpos/products/ui/dto/request/ProductCreateRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.products.ui.dto;
+package kitchenpos.products.ui.dto.request;
 
 import java.math.BigDecimal;
