+++ src/main/java/kitchenpos/products/ui/dto/response/ProductCreateResponse.java
@@ -1,4 +1,6 @@
-package kitchenpos.products.ui.dto;
+package kitchenpos.products.ui.dto.response;
+
+import kitchenpos.products.domain.Product;
 
 import java.math.BigDecimal;
 import java.util.UUID;
@@ -17,8 +19,8 @@ public class ProductCreateResponse {
         this.price = price;
     }
 
-    public static ProductCreateResponse of(UUID id, String name, BigDecimal price) {
-        return new ProductCreateResponse(id, name, price);
+    public static ProductCreateResponse of(Product product) {
+        return new ProductCreateResponse(product.getId(), product.getName().getValue(), product.getPrice().getValue());
     }
 
     public UUID getId() {
