+++ src/main/java/kitchenpos/products/ui/dto/response/ProductResponse.java
@@ -1,4 +1,6 @@
-package kitchenpos.products.ui.dto;
+package kitchenpos.products.ui.dto.response;
+
+import kitchenpos.products.domain.Product;
 
 import java.math.BigDecimal;
 import java.util.UUID;
@@ -17,8 +19,8 @@ public class ProductResponse {
         this.price = price;
     }
 
-    public static ProductResponse of(UUID id, String name, BigDecimal price) {
-        return new ProductResponse(id, name, price);
+    public static ProductResponse of(Product product) {
+        return new ProductResponse(product.getId(), product.getName().getValue(), product.getPrice().getValue());
     }
 
     public UUID getId() {
