+++ src/test/java/kitchenpos/products/application/ProductServiceTest.java
@@ -5,12 +5,12 @@ import kitchenpos.menus.domain.Menu;
 import kitchenpos.menus.domain.MenuRepository;
 import kitchenpos.products.domain.Product;
 import kitchenpos.products.domain.ProductRepository;
-import kitchenpos.products.infra.PurgomalumClient;
-import kitchenpos.products.ui.dto.ProductChangePriceRequest;
-import kitchenpos.products.ui.dto.ProductChangePriceResponse;
-import kitchenpos.products.ui.dto.ProductCreateRequest;
-import kitchenpos.products.ui.dto.ProductCreateResponse;
-import kitchenpos.products.ui.dto.ProductResponse;
+import kitchenpos.products.domain.PurgomalumClient;
+import kitchenpos.products.ui.dto.request.ProductChangePriceRequest;
+import kitchenpos.products.ui.dto.request.ProductCreateRequest;
+import kitchenpos.products.ui.dto.response.ProductChangePriceResponse;
+import kitchenpos.products.ui.dto.response.ProductCreateResponse;
+import kitchenpos.products.ui.dto.response.ProductResponse;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
