+++ src/test/java/kitchenpos/products/domain/ProductTest.java
@@ -1,7 +1,6 @@
 package kitchenpos.products.domain;
 
 import kitchenpos.products.application.FakePurgomalumClient;
-import kitchenpos.products.infra.PurgomalumClient;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
